2022 day01

+++ aoc/transformations.go
@@ -4,20 +4,20 @@ import (
 	"strconv"
 )
 
-func ToIntBase10(text string) int64 {
+func ToIntBase10(text string) int {
 	number, err := strconv.ParseInt(text, 10, 0)
 	if err != nil {
 		panic(err)
 	}
-	return number
+	return int(number)
 }
 
-func ToIntBase2(text string) int64 {
+func ToIntBase2(text string) int {
 	number, err := strconv.ParseInt(text, 2, 0)
 	if err != nil {
 		panic(err)
 	}
-	return number
+	return int(number)
 }
 
 func ToString(text string) string {
